Extracted utilities nav items into separate components
